Implemented the question resource
Added the controller and routes for it. After creating a new question, the user who created it has its questions column updated with the id of the created question.

The BaseController class no longer uses events (because they are async and at some point I will need sync code to modify queries) and instead uses callbacks.

Changed the email column validation function so that it no longer tries to validate the email on user update (if the user _id column is no null).